+++ NEWS.md
@@ -4,7 +4,8 @@
 
 ### Developement version
   
-  * Change `wave` such that we could choose if we would like to impose fixed sample size.
+  * add <arXiv:1910.13152> in DESCRIPTION file.
+  * add `fixedSize` option to choose if we would like to impose fixed sample size. Default is equal `TRUE`
   * add .9000 to show that is the developement version.